+++ micro-modeling/src/main/java/io/event/thinking/micro/modeling/CommandDispatcher.java
@@ -1,0 +1,12 @@
+package io.event.thinking.micro.modeling;
+
+import reactor.core.publisher.Mono;
+
+import java.util.function.Supplier;
+
+public interface CommandDispatcher {
+
+    <T> Mono<Long> dispatch(T cmd);
+
+    <T> void register(Class<T> commandType, Supplier<CommandModel<T>> model);
+}
