+++ micro-modeling/src/main/java/io/event/thinking/micro/modeling/Event.java
@@ -1,0 +1,16 @@
+package io.event.thinking.micro.modeling;
+
+import io.event.thinking.eventstore.Tag;
+
+import java.util.Set;
+
+public record Event(Set<Tag> tags, Object payload) {
+
+    public static Event event(Set<Tag> tags, Object payload) {
+        return new Event(tags, payload);
+    }
+
+    public static Event event(Object payload, Tag... tags) {
+        return new Event(Set.of(tags), payload);
+    }
+}
