+++ micro-modeling/src/main/java/io/event/thinking/micro/modeling/Serializer.java
@@ -1,0 +1,31 @@
+package io.event.thinking.micro.modeling;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
+
+public interface Serializer {
+
+    default byte[] serialize(Object obj) {
+        var bos = new ByteArrayOutputStream();
+
+        try (var out = new ObjectOutputStream(bos)) {
+            out.writeObject(obj);
+            out.flush();
+            return bos.toByteArray();
+        } catch (Exception ex) {
+            throw new RuntimeException(ex);
+        }
+    }
+
+    default Object deserialize(byte[] bytes) {
+        var bis = new ByteArrayInputStream(bytes);
+
+        try (var in = new ObjectInputStream(bis)) {
+            return in.readObject();
+        } catch (Exception ex) {
+            throw new RuntimeException(ex);
+        }
+    }
+}
