+++ sample-faculty/src/main/java/io/event/thinking/sample/faculty/api/command/SubscribeStudent.java
@@ -1,0 +1,5 @@
+package io.event.thinking.sample.faculty.api.command;
+
+public record SubscribeStudent(String studentId, String courseId) {
+
+}
