+++ sample-faculty/src/main/java/io/event/thinking/sample/faculty/api/command/UnsubscribeStudent.java
@@ -1,0 +1,5 @@
+package io.event.thinking.sample.faculty.api.command;
+
+public record UnsubscribeStudent(String studentId, String courseId) {
+
+}
