+++ sample-faculty/src/main/java/io/event/thinking/sample/faculty/api/event/CourseCapacityChanged.java
@@ -1,0 +1,8 @@
+package io.event.thinking.sample.faculty.api.event;
+
+import java.io.Serializable;
+
+public record CourseCapacityChanged(String id, int capacity) implements Serializable {
+
+    public static final String NAME = "CourseCapacityChanged";
+}
