+++ sample-faculty/src/main/java/io/event/thinking/sample/faculty/api/event/CourseCreated.java
@@ -1,0 +1,8 @@
+package io.event.thinking.sample.faculty.api.event;
+
+import java.io.Serializable;
+
+public record CourseCreated(String id, int capacity) implements Serializable {
+
+    public static final String NAME = "CourseCreated";
+}
