+++ sample-faculty/src/main/java/io/event/thinking/sample/faculty/api/event/StudentEnrolledFaculty.java
@@ -1,0 +1,8 @@
+package io.event.thinking.sample.faculty.api.event;
+
+import java.io.Serializable;
+
+public record StudentEnrolledFaculty(String id, String firstName, String lastName) implements Serializable {
+
+    public static final String NAME = "StudentEnrolled";
+}
