+++ sample-faculty/src/main/java/io/event/thinking/sample/faculty/api/event/StudentSubscribed.java
@@ -1,0 +1,8 @@
+package io.event.thinking.sample.faculty.api.event;
+
+import java.io.Serializable;
+
+public record StudentSubscribed(String studentId, String courseId) implements Serializable {
+
+    public static final String NAME = "StudentSubscribed";
+}
