+++ sample-faculty/src/main/java/io/event/thinking/sample/faculty/api/event/StudentUnsubscribed.java
@@ -1,0 +1,9 @@
+package io.event.thinking.sample.faculty.api.event;
+
+import java.io.Serializable;
+
+public record StudentUnsubscribed(String studentId, String courseId) implements Serializable {
+
+    public static final String NAME = "StudentUnsubscribed";
+
+}
