+++ sample-faculty/src/main/java/io/event/thinking/sample/faculty/model/Constants.java
@@ -1,0 +1,7 @@
+package io.event.thinking.sample.faculty.model;
+
+public class Constants {
+
+    public static final String STUDENT_ID = "studentId";
+    public static final String COURSE_ID = "courseId";
+}
