--- src/test/java/io/event/thinking/eventstore/sample/api/command/SubscribeStudent.java
@@ -1,5 +1,0 @@
-package io.event.thinking.eventstore.sample.api.command;
-
-public record SubscribeStudent(String studentId, String courseId) {
-
-}
