--- src/test/java/io/event/thinking/eventstore/sample/api/command/UnsubscribeStudent.java
@@ -1,5 +1,0 @@
-package io.event.thinking.eventstore.sample.api.command;
-
-public record UnsubscribeStudent(String studentId, String courseId) {
-
-}
