--- src/test/java/io/event/thinking/eventstore/sample/api/event/CourseCapacityChanged.java
@@ -1,8 +1,0 @@
-package io.event.thinking.eventstore.sample.api.event;
-
-import java.io.Serializable;
-
-public record CourseCapacityChanged(String id, int capacity) implements Serializable {
-
-    public static final String NAME = "CourseCapacityChanged";
-}
