--- src/test/java/io/event/thinking/eventstore/sample/api/event/StudentEnrolledFaculty.java
@@ -1,8 +1,0 @@
-package io.event.thinking.eventstore.sample.api.event;
-
-import java.io.Serializable;
-
-public record StudentEnrolledFaculty(String id, String firstName, String lastName) implements Serializable {
-
-    public static final String NAME = "StudentEnrolled";
-}
