--- src/test/java/io/event/thinking/eventstore/sample/api/event/StudentSubscribed.java
@@ -1,8 +1,0 @@
-package io.event.thinking.eventstore.sample.api.event;
-
-import java.io.Serializable;
-
-public record StudentSubscribed(String studentId, String courseId) implements Serializable {
-
-    public static final String NAME = "StudentSubscribed";
-}
