--- src/test/java/io/event/thinking/eventstore/sample/api/event/StudentUnsubscribed.java
@@ -1,9 +1,0 @@
-package io.event.thinking.eventstore.sample.api.event;
-
-import java.io.Serializable;
-
-public record StudentUnsubscribed(String studentId, String courseId) implements Serializable {
-
-    public static final String NAME = "StudentUnsubscribed";
-
-}
