--- src/test/java/io/event/thinking/eventstore/sample/microdcb/Serializer.java
@@ -1,31 +1,0 @@
-package io.event.thinking.eventstore.sample.microdcb;
-
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-
-public interface Serializer {
-
-    default byte[] serialize(Object obj) {
-        var bos = new ByteArrayOutputStream();
-
-        try (var out = new ObjectOutputStream(bos)) {
-            out.writeObject(obj);
-            out.flush();
-            return bos.toByteArray();
-        } catch (Exception ex) {
-            throw new RuntimeException(ex);
-        }
-    }
-
-    default Object deserialize(byte[] bytes) {
-        var bis = new ByteArrayInputStream(bytes);
-
-        try (var in = new ObjectInputStream(bis)) {
-            return in.readObject();
-        } catch (Exception ex) {
-            throw new RuntimeException(ex);
-        }
-    }
-}
