--- src/test/java/io/event/thinking/eventstore/sample/model/Constants.java
@@ -1,7 +1,0 @@
-package io.event.thinking.eventstore.sample.model;
-
-public class Constants {
-
-    public static final String STUDENT_ID = "studentId";
-    public static final String COURSE_ID = "courseId";
-}
