+++ store/src/main/java/io/event/thinking/eventstore/ConsistencyCondition.java
