+++ store/src/main/java/io/event/thinking/eventstore/Criteria.java
