+++ store/src/main/java/io/event/thinking/eventstore/Criterion.java
