+++ store/src/main/java/io/event/thinking/eventstore/Event.java
