+++ store/src/main/java/io/event/thinking/eventstore/InvalidConsistencyConditionException.java
