+++ store/src/main/java/io/event/thinking/eventstore/MarkedEvents.java
