+++ store/src/main/java/io/event/thinking/eventstore/SequencedEvent.java
