+++ store/src/main/java/io/event/thinking/eventstore/Tag.java
