+++ store/src/main/java/io/event/thinking/eventstore/inmemory/InMemoryEventStore.java
