+++ store/src/test/java/io/event/thinking/eventstore/CriteriaTest.java
