+++ store/src/test/java/io/event/thinking/eventstore/inmemory/InMemoryEventStoreUseCaseTest.java
